# Optimize local build and test speed