--- a/build.gradle
+++ b/build.gradle
@@ -125,6 +125,9 @@ subprojects { project ->
       exceptionFormat = 'full'
     }
   }
+  tasks.withType(Test) {
+    maxParallelForks = (int) (Runtime.runtime.availableProcessors().intdiv(2) ?: 1)
+  }
 
   if (platform == "jdk8alpn") {
     // Add alpn-boot on Java 8 so we can use HTTP/2 without a stable API.