--- a/gradle.properties
+++ b/gradle.properties
@@ -14,3 +14,6 @@ POM_LICENCE_DIST=repo
 
 POM_DEVELOPER_ID=square
 POM_DEVELOPER_NAME=Square, Inc.
+
+org.gradle.parallel=true
+org.gradle.configureondemand=true