@yschimke 2019-05-07
There is overlap with https://github.com/square/okhttp/pull/4902

We've seen flaky builds, presumably based on load/delays/timeouts etc.  Obviously these differ between Travis, CircleCI and each developers setup.

I'd encourage stress testing this change e.g. setting a stupidly high concurrency and working back to a sane level.  If you can run at 4x (instead of 0.5x) we can be confident this is fine.