@TangHuaiZhe 2019-05-08
> There is overlap with #4902
> 
> We've seen flaky builds, presumably based on load/delays/timeouts etc. Obviously these differ between Travis, CircleCI and each developers setup.
> 
> I'd encourage stress testing this change e.g. setting a stupidly high concurrency and working back to a sane level. If you can run at 4x (instead of 0.5x) we can be confident this is fine.


I tested 8 times in the same code of the same project, and adjusted `maxParallelForks` to 4 times、2 times 、1 time 、0.5 times 、4 times 、2 times、1 time、0.5 times.
Every time execute:
```
./gradlew clean
./gradlew build --scan
```
On my dual-core, four-threaded machine, 
maxParallelForks = 2 at 0.5x ; maxParallelForks = 4 at 1x;
when maxParallelForks>=4  the execution speed is almost the same but all faster than 0.5 times ,and all test passed.

I'll close the pull request because still worried about the side effects on continuous-integration and Sonatype as described by @JakeWharton .

And attach the code comment of `maxParallelForks` in the gradle source code:
`maxParallelForks` — default: 1::
You can run your tests in parallel by setting this property to a value greater than 1. This may make your test suites complete faster, particularly if you run them on a multi-core CPU. **When using parallel test execution, make sure your tests are properly isolated from one another. Tests that interact with the filesystem are particularly prone to conflict, causing intermittent test failures**.

Test reports are as follows:
4*availableProcessors:
Total time 2m 31.166s
Test  time 1m 2.437s
https://scans.gradle.com/s/xuo66z56p2dru/

2*availableProcessors:
Total time 2m 15.016s
Test  time 58.165s
https://scans.gradle.com/s/yb5pwzhfez3sy/

1*availableProcessors:
Total time 1m 51.434s
Test  time 1m 3.079s
https://scans.gradle.com/s/3lxntiqg2ncxk/

0.5*availableProcessors:
Total time 2m 12.951s
Test  time 1m 16.367s
https://scans.gradle.com/s/zmdjczetwiqjq/

4*availableProcessors:
Total time 2m 19.476s
Test  time 1m 0.637s
https://scans.gradle.com/s/2s2jwmek6b4py

2*availableProcessors:
Total time 2m 12.848s
Test  time 1m 2.744s
https://scans.gradle.com/s/qgsb5juwp2o4u


1*availableProcessors:
Total time 2m 23.956s
Test  time 1m 6.671s
https://scans.gradle.com/s/g5mp4abfkt4ju


0.5*availableProcessors:
Total time 2m 21.340s
Test  time 1m 21.020s
https://scans.gradle.com/s/nn7ag6i7eapog


There is a way to open gradle parallel compilation only locally but  no effect to test parallel execution.
Create a  file  `gradle.properties` in the `~/.gradle/` directory, plus the following:
```
org.gradle.daemon=true
org.gradle.parallel=true
org.gradle.configureondemand=true
```

Thanks.